+++ crates/rust-client/src/test_utils/common.rs
@@ -430,11 +430,10 @@ pub async fn assert_account_has_single_asset(
     asset_account_id: AccountId,
     expected_amount: u64,
 ) {
-    let regular_account: Account =
-        client.get_account(account_id).await.unwrap().unwrap().try_into().unwrap();
+    let assets = client.new_account_reader(account_id).assets().await.unwrap();
 
-    assert_eq!(regular_account.vault().assets().count(), 1);
-    let asset = regular_account.vault().assets().next().unwrap();
+    assert_eq!(assets.len(), 1);
+    let asset = &assets[0];
 
     if let Asset::Fungible(fungible_asset) = asset {
         assert_eq!(fungible_asset.faucet_id(), asset_account_id);