+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ### Changes
 
-* Added a `StorageReader` accessible through `Client::new_storage_reader` to read storage slots without having to load the whole account ([#1713](https://github.com/0xMiden/miden-client/pull/1713)). 
+* [BREAKING] Added a `AccountReader` accessible through `Client::new_account_reader` to read account data without needing to load the whole `Account` ([#1716](https://github.com/0xMiden/miden-client/pull/1716)). 
 
 
 ## 0.13.0 (2026-01-28)